Replaced double with decimal in calculation. Fixed #2 partially

+++ CSharpPICollision/AuxiliaryObjects.cs
@@ -5,25 +5,25 @@
     /// </summary>
     /// <param name="Point1">Segment start</param>
     /// <param name="Point2">Segment end</param>
-    readonly record struct Segment(double Point1, double Point2);
+    readonly record struct Segment(decimal Point1, decimal Point2);
     /// <summary>
     /// Represents time to collision of two <seealso cref="PhysicalObject"/>
     /// </summary>
     /// <param name="Object1">First interacting object</param>
     /// <param name="Object2">Second interacting object</param>
     /// <param name="Time">Time to collision</param>
-    readonly record struct Collision(PhysicalObject Object1, PhysicalObject Object2, double Time);
+    readonly record struct Collision(PhysicalObject Object1, PhysicalObject Object2, decimal Time);
     /// <summary>
     /// Represents <seealso cref="PhysicalObject"/> properties change after interaction
     /// </summary>
     /// <param name="Position">Position after collision</param>
     /// <param name="Velocity">Velocity after collision</param>    
-    readonly record struct CollisionResponse(double Position, double Velocity);
+    readonly record struct CollisionResponse(decimal Position, decimal Velocity);
     /// <summary>
     /// Represents horizontal plane properties
     /// </summary>
     /// <param name="Y">Y coordinate of axis</param>
-    public record class HorizontalAxis(double Y);
+    public record class HorizontalAxis(decimal Y);
 
     /// <summary>
     /// Pepresent target of <see cref="CollisionResponse"/>>
